Migrate mgmtsystem_nonconformity_analytic
