--- __unported__/mgmtsystem_nonconformity_analytic/__init__.py
@@ -1,3 +1,0 @@
-# -*- encoding: utf-8 -*-
-
-from . import mgmtsystem_nonconformity
