+++ mgmtsystem_nonconformity_analytic/README.rst
@@ -1,0 +1,54 @@
+.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
+    :alt: License: AGPL-3
+
+Nonconformity - Analytic Account
+================================
+
+This module was written to extend the nonconformity to allow you to create a project instead of actions.
+
+Usage
+=====
+
+To use this module, you need to:
+
+* create a nonconformity
+* walk through the different steps of the workflow to create your action plan
+* create a project instead of actions to correct or prevent the nonconformity
+
+Known issues / Roadmap
+======================
+
+* None
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/management-system/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
+`here <https://github.com/OCA/management-system/issues/new?body=module:%20mgmtsystem_nonconformity_analytic%0Aversion:%208.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Credits
+=======
+
+Contributors
+------------
+
+* Daniel Reis <dreis.pt@gmail.com>
+* Loïc Lacroix <loic.lacroix@savoirfairelinux.com>
+* Maxime Chambreuil <maxime.chambreuil@savoirfairelinux.com>
+
+Maintainer
+----------
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+This module is maintained by the OCA.
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+To contribute to this module, please visit http://odoo-community.org.
